Move logger tests to their own directory
