+++ tests/Logger/AbstractLoggerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests;
+namespace Tests\Logger;
 
 use PHPUnit\Framework\TestCase;
